requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
numpy==1.26.2
pandas==2.1.3
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
